+++ android/src/main/kotlin/com/mapbox/maps/mapbox_maps/LocationComponentController.kt
@@ -8,8 +8,10 @@ import com.mapbox.maps.mapbox_maps.pigeons.*
 import com.mapbox.maps.plugin.locationcomponent.createDefault2DPuck
 import com.mapbox.maps.plugin.locationcomponent.location
 
-class LocationComponentController(private val mapView: MapView, private val context: Context) :
-  _LocationComponentSettingsInterface {
+class LocationComponentController(
+  private val mapView: MapView,
+  private val context: Context
+) : _LocationComponentSettingsInterface {
   override fun getSettings(): LocationComponentSettings = mapView.location.toFLT(context)
 
   override fun updateSettings(settings: LocationComponentSettings, useDefaultPuck2DIfNeeded: Boolean) {