+++ lib/src/style/layer/line_layer.dart
@@ -256,7 +256,7 @@ class LineLayer extends Layer {
   List<Object>? lineWidthExpression;
 
   @override
-  String _encode() {
+  Future<String> _encode() async {
     var layout = {};
     if (visibilityExpression != null) {
       layout["visibility"] = visibilityExpression!;