+++ lib/src/turf_adapters.dart
@@ -35,7 +35,8 @@ final class Polygon extends turf.Polygon {
     return Polygon(bbox: polygon.bbox, coordinates: polygon.coordinates);
   }
 
-    Polygon.fromPoints({turf.BBox? bbox, required List<List<Point>> points}) : super.fromPoints(bbox: bbox, points: points);
+  Polygon.fromPoints({turf.BBox? bbox, required List<List<Point>> points})
+      : super.fromPoints(bbox: bbox, points: points);
 }
 
 final class LineString extends turf.LineString {
@@ -55,5 +56,6 @@ final class LineString extends turf.LineString {
     return LineString(bbox: line.bbox, coordinates: line.coordinates);
   }
 
-  LineString.fromPoints({turf.BBox? bbox, required List<Point> points}) : super.fromPoints(bbox: bbox, points: points);
+  LineString.fromPoints({turf.BBox? bbox, required List<Point> points})
+      : super.fromPoints(bbox: bbox, points: points);
 }